chore: colocate routine move CLI adapter
