+++ .changeset/move-routine-cli-route.md
@@ -1,0 +1,5 @@
+---
+"moadim": patch
+---
+
+Move the routine move CLI adapter next to the REST and MCP move route modules without changing behavior.
